# gatemint/daemon/kmd/api/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "API for KMD (Key Management Daemon)",
    "title": "for KMD HTTP API",
    "contact": {
      "email": "contact@algorand.com"
    },
    "version": "0.0.1"
  },
  "host": "localhost",
  "basePath": "/",
  "paths": {
    "/swagger.json": {
      "get": {
        "description": "Returns the entire swagger spec in json.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "summary": "Gets the current swagger spec.",
        "operationId": "SwaggerHandler",
        "responses": {
          "200": {
            "description": "The current swagger spec",
            "schema": {
              "type": "string"
            }
          },
          "default": {
            "description": "Unknown Error"
          }
        }
      }
    },
...
    "/versions": {
      "get": {
        "produces": [
          "application/json"
        ],
        "summary": "Retrieves the current version",
        "operationId": "GetVersion",
        "parameters": [
          {
            "name": "Versions Request",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/VersionsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/VersionsResponse"
          }
        }
      }
    }
  },
  "definitions": {
...
    "VersionsRequest": {
      "description": "VersionsRequest is the request for `GET /versions`",
      "type": "object",
      "x-go-package": "github.com/gatechain/gatemint/daemon/kmd/lib/kmdapi"
    },
    "VersionsResponse": {
      "description": "VersionsResponse is the response to `GET /versions`\nfriendly:VersionsResponse",
      "type": "object",
      "properties": {
        "versions": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Versions"
        }
      },
      "x-go-package": "github.com/gatechain/gatemint/daemon/kmd/lib/kmdapi"
    },
...
  },
  "responses": {
...
    "VersionsResponse": {
      "description": "SwaggerVersionsResponse is an autogenerated wrapper from genSwaggerWrappers.py",
      "schema": {
        "$ref": "#/definitions/VersionsResponse"
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "description": "Generated header parameter. This value can be found in `/kmd/data/dir/kmd.token`. Example value: '330b2e4fc9b20f4f89812cf87f1dabeb716d23e3f11aec97a61ff5f750563b78'",
      "type": "apiKey",
      "name": "X-KMD-API-Token",
      "in": "header",
      "x-example": "330b2e4fc9b20f4f89812cf87f1dabeb716d23e3f11aec97a61ff5f750563b78"
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}